- Large refactor to remove string as a basic type and move it to an ExternalReflector. - Added initial support for Double basic type. Needs work. - Added a constant pool for strings. There should be less copying of strings because of this. - The default string type can now be configured in Configuration.h. - Added LLVMPreGeneratedExpression which can pass an llvm::Value through a CatTypeExpression during code generation. - Added /bigobj flag to build configuration for MSVC. - Fixed issue with temporaries being deleted prematurely. - ExpressionAssignAny will now call operator= when applicable.

+++ include/jitcat/InfixOperatorOptimizer.h
@@ -44,7 +44,7 @@ namespace jitcat::AST
 		static CatTypedExpression* tryCollapseLogicalAnd(std::unique_ptr<CatTypedExpression>& lhs, std::unique_ptr<CatTypedExpression>& rhs);
 		static CatTypedExpression* tryCollapseLogicalOr(std::unique_ptr<CatTypedExpression>& lhs, std::unique_ptr<CatTypedExpression>& rhs);
 
-		static bool typedExpressionEqualsConstant(CatTypedExpression* expression, float constant);
+		static bool typedExpressionEqualsConstant(CatTypedExpression* expression, double constant);
 		static bool typedExpressionEqualsConstant(CatTypedExpression* expression, bool constant);
 	};
